misc: move esm-utils.js to shared/
