+++ core/scripts/i18n/collect-strings.js
@@ -23,7 +23,7 @@ import {collectAndBakeCtcStrings} from './bake-ctc-to-lhl.js';
 import {pruneObsoleteLhlMessages} from './prune-obsolete-lhl-messages.js';
 import {countTranslatedMessages} from './count-translated.js';
 import {LH_ROOT} from '../../../root.js';
-import {resolveModulePath} from '../../../esm-utils.js';
+import {resolveModulePath} from '../../../shared/esm-utils.js';
 
 // Match declarations of UIStrings, terminating in either a `};\n` (very likely to always be right)
 // or `}\n\n` (allowing semicolon to be optional, but insisting on a double newline so that an
