+++ tsconfig.json
@@ -12,7 +12,6 @@
   ],
   "include": [
     "root.js",
-    "esm-utils.js",
     "cli/**/*.js",
     "core/**/*.js",
     "core/index.cjs",
